Refactor data model types and enhance command handling
- Changed ColumnId and RowId types from long to int in ColumnEntity and RowEntity.
- Added GetOrThrow method to DictionaryExtensions for safer dictionary access.
- Updated DataGenerationSchemaViewModel to remove unused undo/redo properties.
- Improved TableStoreDbContext to include additional SQL commands for database setup.
- Enhanced command handlers to support loading state management.
- Introduced IIdentifiable and ISerialisationData interfaces for better data handling.
- Updated UI bindings in TableDataView.axaml for improved data display.

+++ TableConverter.Utilities/Database/Models/CellEntity.cs
@@ -5,15 +5,15 @@ namespace TableConverter.Utilities.Database.Models;
 
 public sealed class CellEntity : INotifyPropertyChanged
 {
-    private long _RowId;
-    public long RowId
+    private int _RowId;
+    public int RowId
     {
         get => _RowId;
         set => SetField(ref _RowId, value);
     }
 
-    private long _ColumnId;
-    public long ColumnId
+    private int _ColumnId;
+    public int ColumnId
     {
         get => _ColumnId;
         set => SetField(ref _ColumnId, value);
